requirements.txt
absl-py==1.0.0
attrs==21.4.0
autopep8==1.6.0
bsuite==0.3.5
chex==0.1.0
click==8.0.4
cloudpickle==2.0.0
cycler==0.11.0
decorator==5.1.1
Deprecated==1.2.13
descartes==1.1.0
distrax==0.1.0
dm-env==1.5
dm-haiku==0.0.6
dm-tree==0.1.6
filelock==3.6.0
flatbuffers==2.0
fonttools==4.29.1
frozendict==2.3.0
gast==0.5.3
grpcio==1.44.0
gym==0.22.0
gym-notices==0.0.4
imageio==2.16.0
importlib-metadata==4.11.1
jax==0.2.21
jaxlib==0.3.0+cuda11.cudnn82
jmp==0.0.2
jsonschema==4.4.0
kiwisolver==1.3.2
matplotlib==3.5.1
mizani==0.7.3
msgpack==1.0.3
mypy==0.931
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.22.2
opt-einsum==3.3.0
optax==0.1.1
packaging==21.3
palettable==3.3.0
pandas==1.4.1
patsy==0.5.2
Pillow==9.0.1
plotnine==0.8.0
protobuf==3.19.4
pycodestyle==2.8.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
ray==1.10.0
redis==4.1.4
rlax==0.1.2
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
statsmodels==0.13.2
tabulate==0.8.9
tensorflow-probability==0.16.0
termcolor==1.1.0
tifffile==2022.2.9
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tqdm==4.62.3
typing_extensions==4.1.1
wrapt==1.13.3
zipp==3.7.0
